worked on stored xss,also added navbar page and a common logout.php for all pages

+++ admin/delete_product.php
@@ -1,8 +1,15 @@
 <?php
+session_start();
+
+if (!isset($_SESSION['id']) || $_SESSION['role'] != 'admin') {
+    header("Location: ../login.php");
+    exit();
+}
 
 include("../db.php");
 
-$id=$_GET['id'];
+$id = intval($_GET['id']);
+
 
 mysqli_query($conn,
 "DELETE FROM products
